Written UI using PyQt6
Hi there,
I had seen this project through reddit and tried to rewrite the UI from scratch in PyQt also, I wrote the code for mica in the ui itself for reducing dependencies. The ui which I wrote in PyQt is slightly different from the original one but there are several advantages of using it than tkinter.
1) Using multi-threading with QThread, we can reduce lag when we enter long passwords.
2) Animations can be added like I added in the progress bar.
3) With the help of darkdetect library, we can add dark mode and light mode support, and while switching themes, ui won't cause many problems as they are seen sometimes with tkinter.

You may review and tell if some other changes are required.